Move trash-specific things to a trait
